Split the operator into three files and also corrected the compilation issue
